+++ pkg/observation/basic_test.go
@@ -1,13 +1,13 @@
-package observations_test
+package observation_test
 
 import (
 	. "github.com/onsi/ginkgo"
-	"iio/pkg/observations"
+	"iio/pkg/observation"
 	"iio/pkg/test"
 )
 
 var _ = Describe("basic", func() {
-	observe := func(observer observations.Observer, cost float64) {
+	observe := func(observer observation.Observer, cost float64) {
 
 	}
 	test.With("should correctly collect all costs and produce a series", func() {
